refactoring: Changes in project structure
