+++ tests/test_lib_cli.py
